Update Despesa.java
Correções do Felipe

+++ src/main/java/com/easyevents/auth_service/domain/model/Despesa.java
@@ -5,6 +5,7 @@ import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Map;
 
@@ -21,7 +22,7 @@ public class Despesa {
     private String id;
     private String titulo;
     private String descricao;
-    private Integer valor;
+    private BigDecimal valor;
     private TipoDespesa tipoDespesa;
 
 }
